+++ protocols/gossipsub/src/protocol.rs
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::{convert::Infallible, pin::Pin};
+use std::{collections::HashMap, convert::Infallible, pin::Pin};
 
 use asynchronous_codec::{Decoder, Encoder, Framed};
 use byteorder::{BigEndian, ByteOrder};
@@ -30,15 +30,10 @@ use libp2p_swarm::StreamProtocol;
 use quick_protobuf::{MessageWrite, Writer};
 
 use crate::{
-    config::{TopicConfigs, ValidationMode},
-    handler::HandlerEvent,
-    rpc_proto::proto,
-    topic::TopicHash,
-    types::{
+    config::{TopicConfigs, ValidationMode}, handler::HandlerEvent, rpc_proto::proto, topic::TopicHash, types::{
         ControlAction, Graft, IDontWant, IHave, IWant, MessageId, PeerInfo, PeerKind, Prune,
         RawMessage, Rpc, Subscription, SubscriptionAction,
-    },
-    ValidationError,
+    }, Config, ValidationError
 };
 
 pub(crate) const SIGNING_PREFIX: &[u8] = b"libp2p-pubsub:";
@@ -67,9 +62,7 @@ pub struct ProtocolConfig {
     /// The Gossipsub protocol id to listen on.
     pub(crate) protocol_ids: Vec<ProtocolId>,
     /// Determines the level of validation to be done on incoming messages.
-    pub(crate) validation_mode: ValidationMode,
-    /// Configurations for topics including max transmit size and mesh parameters.
-    pub(crate) config: TopicConfigs,
+    pub(crate) validation_mode: ValidationMode
 }
 
 impl Default for ProtocolConfig {
@@ -81,7 +74,6 @@ impl Default for ProtocolConfig {
                 GOSSIPSUB_1_1_0_PROTOCOL,
                 GOSSIPSUB_1_0_0_PROTOCOL,
             ],
-            config: TopicConfigs::default(),
         }
     }
 }
@@ -123,9 +115,9 @@ where
             Framed::new(
                 socket,
                 GossipsubCodec::new(
-                    self.config.default_max_transmit_size,
+                    Config::default_max_transmit_size(),
                     self.validation_mode,
-                    self.config,
+                    TopicConfigs::default()
                 ),
             ),
             protocol_id.kind,
@@ -146,9 +138,9 @@ where
             Framed::new(
                 socket,
                 GossipsubCodec::new(
-                    self.config.default_max_transmit_size,
+                    Config::default_max_transmit_size(),
                     self.validation_mode,
-                    self.config,
+                    TopicConfigs::default()
                 ),
             ),
             protocol_id.kind,
@@ -163,21 +155,21 @@ pub struct GossipsubCodec {
     validation_mode: ValidationMode,
     /// The codec to handle common encoding/decoding of protobuf messages
     codec: quick_protobuf_codec::Codec<proto::RPC>,
-    /// Configurations for topics including max transmit size and mesh parameters.
-    config: TopicConfigs,
+    /// Configurations for topics including max transmit size and mesh parameters
+    config: TopicConfigs
 }
 
 impl GossipsubCodec {
     pub fn new(
         max_length: usize,
         validation_mode: ValidationMode,
-        config: TopicConfigs,
+        config: TopicConfigs
     ) -> GossipsubCodec {
         let codec = quick_protobuf_codec::Codec::new(max_length);
         GossipsubCodec {
             validation_mode,
             codec,
-            config,
+            config
         }
     }
 
@@ -275,7 +267,7 @@ impl Decoder for GossipsubCodec {
                     validated: false,
                 };
 
-                invalid_messages.push((message, ValidationError::TransformFailed));
+                invalid_messages.push((message, ValidationError::MessageSizeTooLargeForTopic));
                 continue;
             }