+++ protocols/gossipsub/CHANGELOG.md
@@ -18,7 +18,7 @@
 - Fix messages were not forwarded to floodsub peers.
   See [PR 5908](https://github.com/libp2p/rust-libp2p/pull/5908)
 
-- Adds functions to ConfigBuilder interface allowing modification of mesh-n-*   params and max transmit sizes per topic. Modifies behaviour.rs, GossipsubCodec, ProtocolConfig to use new config defaults and methods to check values against topics.
+- Allow customizing max transmit size and mesh-n-* parameters per topic.
   See [PR 5868](https://github.com/libp2p/rust-libp2p/pull/5868)
 
 ## 0.48.0